+++ test_project/special/sample.ex
@@ -1,0 +1,10 @@
+# Sample Elixir file
+defmodule HelloWorld do
+  alias String.Chars
+  import IO
+  require Logger
+
+  def hello do
+    puts "Hello, Elixir!"
+  end
+end