+++ test_project/special/sample.go
@@ -1,0 +1,11 @@
+// Sample Go file
+package main
+
+import (
+    "fmt"
+    "strings"
+)
+
+func main() {
+    fmt.Println("Hello, Go!")
+}