+++ test_project/special/sample.swift
@@ -1,0 +1,9 @@
+// Sample Swift file
+import Foundation
+import UIKit
+
+class HelloWorld {
+    func sayHello() {
+        print("Hello, Swift!")
+    }
+}