+++ test_project/special/sample.tf
@@ -1,0 +1,17 @@
+# Sample Terraform file
+provider "aws" {
+  region = "us-west-2"
+}
+
+module "vpc" {
+  source = "terraform-aws-modules/vpc/aws"
+  version = "3.14.0"
+  
+  name = "my-vpc"
+  cidr = "10.0.0.0/16"
+}
+
+resource "aws_instance" "web" {
+  ami           = "ami-0c55b159cbfafe1f0"
+  instance_type = "t2.micro"
+}